+++ src/main.rs
@@ -1,15 +1,18 @@
+mod config;
+mod llm;
+
 use anyhow::{anyhow, Context, Result};
 use clap::{ArgGroup, Parser, Subcommand};
-use reqwest::blocking::Client;
-use serde::{Deserialize, Serialize};
-use std::collections::{BTreeMap, HashSet};
-use std::env;
+use config::Config;
+use llm::{LlmClient, NoopClient};
+use llm::openai::OpenAiClient;
+use std::collections::HashSet;
 use std::io::{self, Write};
 use std::process::Command as GitCommand;
 
 /// How the user categorizes each file in interactive mode.
-#[derive(Debug, Clone, Copy, Serialize)]
-enum FileCategory {
+#[derive(Debug, Clone, Copy, serde::Serialize)]
+pub enum FileCategory {
     Main,        // 1
     Supporting,  // 2
     Consequence, // 3
@@ -27,7 +30,7 @@ impl FileCategory {
         }
     }
 
-    fn as_str(&self) -> &'static str {
+    pub fn as_str(&self) -> &'static str {
         match self {
             FileCategory::Main => "main",
             FileCategory::Supporting => "supporting",
@@ -39,22 +42,22 @@ impl FileCategory {
 
 /// Represents a single staged file's change and metadata.
 #[derive(Debug, Clone)]
-struct FileChange {
-    path: String,
-    category: FileCategory,
-    diff: String,
-    summary: Option<String>, // Filled by per-file model call (or dummy)
+pub struct FileChange {
+    pub path: String,
+    pub category: FileCategory,
+    pub diff: String,
+    pub summary: Option<String>, // Filled by per-file model call (or dummy)
 }
 
 /// How we want to summarize a PR.
 #[derive(Debug, Clone, Copy)]
-enum PrSummaryMode {
+pub enum PrSummaryMode {
     ByCommits,
     ByPrs,
 }
 
 impl PrSummaryMode {
-    fn as_str(&self) -> &'static str {
+    pub fn as_str(&self) -> &'static str {
         match self {
             PrSummaryMode::ByCommits => "commits",
             PrSummaryMode::ByPrs => "prs",
@@ -64,11 +67,11 @@ impl PrSummaryMode {
 
 /// A commit involved in the PR range, plus any detected PR number.
 #[derive(Debug, Clone)]
-struct PrItem {
-    commit_hash: String,
-    title: String,
-    body: String,
-    pr_number: Option<u32>,
+pub struct PrItem {
+    pub commit_hash: String,
+    pub title: String,
+    pub body: String,
+    pub pr_number: Option<u32>,
 }
 
 /// CLI options
@@ -83,39 +86,38 @@ struct PrItem {
         .args(["model", "no_model"])
         .multiple(false)
 ))]
-struct Cli {
+pub struct Cli {
     /// Interactive mode: classify each file and do per-file summaries
     #[arg(long, global = true)]
-    ask: bool,
+    pub ask: bool,
 
     /// Debug mode: log prompts, responses, token usage
     #[arg(long, global = true)]
-    debug: bool,
+    pub debug: bool,
 
     /// Model name to use (e.g. gpt-4o-mini). If 'none', acts like --no-model.
     #[arg(long, global = true)]
-    model: Option<String>,
-
+    pub model: Option<String>,
     /// Disable model calls; return dummy responses instead
     #[arg(long, global = true)]
-    no_model: bool,
+    pub no_model: bool,
 
     /// API key (otherwise uses OPENAI_API_KEY env var)
     #[arg(long, env = "OPENAI_API_KEY", global = true)]
-    api_key: Option<String>,
+    pub api_key: Option<String>,
 
     /// Optional: a brief human description of the ticket (for commit/PR summaries)
     #[arg(long, global = true)]
-    ticket_summary: Option<String>,
+    pub ticket_summary: Option<String>,
 
     /// Subcommand (e.g. 'pr')
     #[command(subcommand)]
-    command: Option<Command>,
+    pub command: Option<Command>,
 }
 
 /// Subcommands, e.g. `commitbot pr develop`
 #[derive(Subcommand, Debug)]
-enum Command {
+pub enum Command {
     /// Generate a Pull Request description by summarizing commit or PR messages
     Pr {
         /// Base branch to compare against (e.g. main or develop)
@@ -132,578 +134,6 @@ enum Command {
         #[arg(long = "commit")]
         commit_mode: bool,
     },
-}
-
-/// Trait for talking to an LLM (real or dummy).
-trait LlmClient {
-    /// Generate a per-file summary based on diff + metadata.
-    fn summarize_file(
-        &self,
-        branch: &str,
-        file: &FileChange,
-        ticket_summary: Option<&str>,
-        debug: bool,
-    ) -> Result<String>;
-
-    /// Generate the final commit message from file summaries + metadata.
-    fn generate_commit_message(
-        &self,
-        branch: &str,
-        files: &[FileChange],
-        ticket_summary: Option<&str>,
-        debug: bool,
-    ) -> Result<String>;
-
-    /// Simple mode: commit message from entire diff.
-    fn generate_commit_message_simple(
-        &self,
-        branch: &str,
-        diff: &str,
-        debug: bool,
-    ) -> Result<String>;
-
-    /// PR mode: generate a PR description from commit/PR messages.
-    fn generate_pr_message(
-        &self,
-        base_branch: &str,
-        from_branch: &str,
-        mode: PrSummaryMode,
-        items: &[PrItem],
-        ticket_summary: Option<&str>,
-        debug: bool,
-    ) -> Result<String>;
-}
-
-/// No-op / dummy model client for development with --no-model or model=none.
-struct NoopClient;
-
-impl LlmClient for NoopClient {
-    fn summarize_file(
-        &self,
-        _branch: &str,
-        file: &FileChange,
-        _ticket_summary: Option<&str>,
-        _debug: bool,
-    ) -> Result<String> {
-        Ok(format!(
-            "[DUMMY SUMMARY] {} ({})",
-            file.path,
-            file.category.as_str()
-        ))
-    }
-
-    fn generate_commit_message(
-        &self,
-        branch: &str,
-        files: &[FileChange],
-        ticket_summary: Option<&str>,
-        _debug: bool,
-    ) -> Result<String> {
-        let mut msg = String::new();
-        msg.push_str("Dummy commit message for testing\n\n");
-        msg.push_str(&format!("Branch: {branch}\n"));
-        if let Some(ts) = ticket_summary {
-            msg.push_str(&format!("Ticket: {ts}\n\n"));
-        } else {
-            msg.push('\n');
-        }
-
-        for f in files.iter().filter(|f| !matches!(f.category, FileCategory::Ignored)) {
-            msg.push_str(&format!(
-                "- {} [{}]: {}\n",
-                f.path,
-                f.category.as_str(),
-                f.summary
-                    .as_deref()
-                    .unwrap_or("[no summary; dummy client]")
-            ));
-        }
-
-        Ok(msg)
-    }
-
-    fn generate_commit_message_simple(
-        &self,
-        branch: &str,
-        _diff: &str,
-        _debug: bool,
-    ) -> Result<String> {
-        Ok(format!(
-            "Dummy simple commit message for branch {branch}\n\n(LLM disabled)"
-        ))
-    }
-
-    fn generate_pr_message(
-        &self,
-        base_branch: &str,
-        from_branch: &str,
-        mode: PrSummaryMode,
-        items: &[PrItem],
-        ticket_summary: Option<&str>,
-        _debug: bool,
-    ) -> Result<String> {
-        let mut msg = String::new();
-        msg.push_str("Dummy PR description for testing\n\n");
-        msg.push_str(&format!(
-            "Base branch: {base}\nFeature branch: {from}\nMode: {mode}\n\n",
-            base = base_branch,
-            from = from_branch,
-            mode = mode.as_str()
-        ));
-
-        if let Some(ts) = ticket_summary {
-            msg.push_str(&format!("Ticket summary: {ts}\n\n"));
-        }
-
-        for item in items {
-            msg.push_str(&format!(
-                "- {} {} (PR #{})\n",
-                &item.commit_hash.chars().take(7).collect::<String>(),
-                item.title.trim(),
-                item.pr_number
-                    .map(|n| n.to_string())
-                    .unwrap_or_else(|| "none".into())
-            ));
-        }
-
-        Ok(msg)
-    }
-}
-
-/// Minimal request/response structs for OpenAI Chat Completions API.
-#[derive(Serialize)]
-struct ChatRequest {
-    model: String,
-    messages: Vec<ChatMessage>,
-}
-
-#[derive(Serialize)]
-struct ChatMessage {
-    role: String,
-    content: String,
-}
-
-#[derive(Deserialize)]
-struct ChatResponse {
-    choices: Vec<ChatChoice>,
-    usage: Option<ChatUsage>,
-}
-
-#[derive(Deserialize)]
-struct ChatChoice {
-    message: ChatMessageResponse,
-}
-
-#[derive(Deserialize)]
-struct ChatMessageResponse {
-    content: String,
-}
-
-#[derive(Deserialize)]
-struct ChatUsage {
-    prompt_tokens: u32,
-    completion_tokens: u32,
-    total_tokens: u32,
-}
-
-/// OpenAI-based implementation of LlmClient.
-struct OpenAiClient {
-    client: Client,
-    api_key: String,
-    model: String,
-}
-
-impl OpenAiClient {
-    fn new(api_key: String, model: String) -> Self {
-        OpenAiClient {
-            client: Client::new(),
-            api_key,
-            model,
-        }
-    }
-
-    fn call_chat(&self, req: &ChatRequest, debug: bool) -> Result<(String, Option<ChatUsage>)> {
-        let url = "https://api.openai.com/v1/chat/completions";
-
-        if debug {
-            eprintln!("[DEBUG] Calling OpenAI model: {}", req.model);
-        }
-
-        let resp = self
-            .client
-            .post(url)
-            .bearer_auth(&self.api_key)
-            .json(req)
-            .send()
-            .context("failed to send request to OpenAI")?;
-
-        if !resp.status().is_success() {
-            let status = resp.status();
-            let text = resp.text().unwrap_or_default();
-            return Err(anyhow!(
-                "OpenAI API error: HTTP {} - {}",
-                status.as_u16(),
-                text
-            ));
-        }
-
-        let chat_resp: ChatResponse = resp.json().context("failed to parse OpenAI response")?;
-        let content = chat_resp
-            .choices
-            .get(0)
-            .map(|c| c.message.content.clone())
-            .ok_or_else(|| anyhow!("no choices returned from OpenAI"))?;
-
-        if debug {
-            if let Some(usage) = &chat_resp.usage {
-                eprintln!(
-                    "[DEBUG] Token usage: prompt={}, completion={}, total={}",
-                    usage.prompt_tokens, usage.completion_tokens, usage.total_tokens
-                );
-            }
-        }
-
-        Ok((content, chat_resp.usage))
-    }
-}
-
-impl LlmClient for OpenAiClient {
-    fn summarize_file(
-        &self,
-        branch: &str,
-        file: &FileChange,
-        ticket_summary: Option<&str>,
-        debug: bool,
-    ) -> Result<String> {
-        let mut system_instructions = String::from(
-            "You are a helpful assistant that explains code changes file-by-file \
-             to later help generate a Git commit message.\n\
-             Focus on intent, not line-by-line diffs.\n\
-             Keep the summary to 2-4 bullet points.",
-        );
-
-        if let Some(ts) = ticket_summary {
-            system_instructions.push_str("\nOverall ticket goal: ");
-            system_instructions.push_str(ts);
-        }
-
-        let user_prompt = format!(
-            "Branch: {branch}\n\
-             File: {path}\n\
-             Category: {category}\n\n\
-             Diff:\n\
-             ```diff\n{diff}\n```",
-            branch = branch,
-            path = file.path,
-            category = file.category.as_str(),
-            diff = file.diff
-        );
-
-        if debug {
-            eprintln!(
-                "[DEBUG] Per-file summarize prompt for {} ({:?}):\n{}",
-                file.path,
-                file.category,
-                truncate(&user_prompt, 2000)
-            );
-        }
-
-        let req = ChatRequest {
-            model: self.model.clone(),
-            messages: vec![
-                ChatMessage {
-                    role: "system".into(),
-                    content: system_instructions,
-                },
-                ChatMessage {
-                    role: "user".into(),
-                    content: user_prompt,
-                },
-            ],
-        };
-
-        let (content, _usage) = self.call_chat(&req, debug)?;
-        Ok(content)
-    }
-
-    fn generate_commit_message(
-        &self,
-        branch: &str,
-        files: &[FileChange],
-        ticket_summary: Option<&str>,
-        debug: bool,
-    ) -> Result<String> {
-        let mut per_file_block = String::new();
-        for file in files.iter().filter(|f| !matches!(f.category, FileCategory::Ignored)) {
-            per_file_block.push_str(&format!(
-                "File: {path}\nCategory: {category}\nSummary:\n{summary}\n\n",
-                path = file.path,
-                category = file.category.as_str(),
-                summary = file
-                    .summary
-                    .as_deref()
-                    .unwrap_or("[missing per-file summary]")
-            ));
-        }
-
-        let mut system_instructions = String::from(
-            "You are a Git commit message assistant.\n\
-             Write a descriptive Git commit message based on the file summaries.\n\
-             Rules:\n\
-             1. Start with a summary line under 50 characters, no formatting.\n\
-             2. Follow with a detailed breakdown grouped by type of change.\n\
-             3. Use headlines (## Migrations, ## Factories, ## Models, etc.).\n\
-             4. Use bullet points under each group.\n\
-             5. If something is new, call it 'Introduced', not 'Refactored'.\n\
-             6. If it fixes broken or incomplete behavior, prefer 'Fixed' or 'Refined'.\n\
-             7. Do not call something a refactor if it is being introduced.\n\
-             8. Avoid generic terms like 'update' or 'improve' unless strictly accurate.\n\
-             9. Group repetitive changes (like renames) instead of repeating them per file.\n\
-             10. Focus on the main purpose and supporting work; only briefly mention consequences.",
-        );
-
-        if let Some(ts) = ticket_summary {
-            system_instructions.push_str("\nOverall ticket goal: ");
-            system_instructions.push_str(ts);
-        }
-
-        let user_prompt = format!(
-            "Branch: {branch}\n\nPer-file summaries:\n\n{per_file}",
-            branch = branch,
-            per_file = per_file_block
-        );
-
-        if debug {
-            eprintln!(
-                "[DEBUG] Final commit-message prompt:\n{}",
-                truncate(&user_prompt, 3000)
-            );
-        }
-
-        let req = ChatRequest {
-            model: self.model.clone(),
-            messages: vec![
-                ChatMessage {
-                    role: "system".into(),
-                    content: system_instructions,
-                },
-                ChatMessage {
-                    role: "user".into(),
-                    content: user_prompt,
-                },
-            ],
-        };
-
-        let (content, _usage) = self.call_chat(&req, debug)?;
-        Ok(content)
-    }
-
-    fn generate_commit_message_simple(
-        &self,
-        branch: &str,
-        diff: &str,
-        debug: bool,
-    ) -> Result<String> {
-        let system_prompt = String::from(
-            "You are a Git commit message assistant.\n\
-             Write a descriptive Git commit message for the given diff.\n\
-             Follow these rules:\n\
-             1. Start with a summary line under 50 characters, no formatting.\n\
-             2. Follow with a detailed breakdown grouped by type of change.\n\
-             3. Use headlines (## Migrations, ## Factories, ## Models, etc.).\n\
-             4. Use bullet points under each group.\n\
-             5. If something is new, call it 'Introduced', not 'Refactored'.\n\
-             6. If it fixes broken or incomplete behavior, prefer 'Fixed' or 'Refined'.\n\
-             7. Do not call something a refactor if it is being introduced.\n\
-             8. Avoid generic terms like 'update' or 'improve' unless strictly accurate.\n\
-             9. Group repetitive changes (like renames) instead of repeating them per file.\n\
-             10. Infer intent where possible from names and context.",
-        );
-
-        let user_prompt = format!(
-            "Branch: {branch}\n\nDiff:\n```diff\n{diff}\n```",
-            branch = branch,
-            diff = diff
-        );
-
-        if debug {
-            eprintln!(
-                "[DEBUG] Simple commit-message prompt:\n{}",
-                truncate(&user_prompt, 3000)
-            );
-        }
-
-        let req = ChatRequest {
-            model: self.model.clone(),
-            messages: vec![
-                ChatMessage {
-                    role: "system".into(),
-                    content: system_prompt,
-                },
-                ChatMessage {
-                    role: "user".into(),
-                    content: user_prompt,
-                },
-            ],
-        };
-
-        let (content, _usage) = self.call_chat(&req, debug)?;
-        Ok(content)
-    }
-
-    fn generate_pr_message(
-        &self,
-        base_branch: &str,
-        from_branch: &str,
-        mode: PrSummaryMode,
-        items: &[PrItem],
-        ticket_summary: Option<&str>,
-        debug: bool,
-    ) -> Result<String> {
-        let mut system_instructions = String::from(
-            "You are a GitHub Pull Request description assistant.\n\
-             Your job is to summarize the *overall goal* of the branch and the important changes.\n\
-             Rules:\n\
-             1. Start with a concise PR title (<= 72 characters, no formatting).\n\
-             2. Then include sections, for example:\n\
-                - ## Overview\n\
-                - ## Changes\n\
-                - ## Testing / Validation\n\
-                - ## Notes / Risks\n\
-             3. Focus on user-visible behavior and domain-level intent, not line-by-line diffs.\n\
-             4. De-emphasize purely mechanical changes (formatting-only, CI-only, or style-only).\n\
-             5. If PR numbers are provided, reference them in the summary (e.g. 'PR #123').\n\
-             6. When multiple PRs contributed, explain how they fit together into a single story.\n\
-             7. Avoid generic phrases like 'misc changes' or 'small fixes'; be specific.",
-        );
-
-        if let Some(ts) = ticket_summary {
-            system_instructions.push_str("\nOverall ticket goal: ");
-            system_instructions.push_str(ts);
-        }
-
-        let mut user_prompt = String::new();
-        user_prompt.push_str(&format!(
-            "Base branch: {base}\nFeature branch: {from}\nSummary mode: {mode}\n\n",
-            base = base_branch,
-            from = from_branch,
-            mode = mode.as_str()
-        ));
-
-        match mode {
-            PrSummaryMode::ByCommits => {
-                user_prompt.push_str("Commit history (oldest first):\n");
-                for item in items {
-                    let short = item.commit_hash.chars().take(7).collect::<String>();
-                    let pr_tag = item
-                        .pr_number
-                        .map(|n| format!(" (PR #{n})"))
-                        .unwrap_or_default();
-                    user_prompt.push_str(&format!(
-                        "- {short}{pr_tag}: {title}\n",
-                        title = item.title.trim()
-                    ));
-                    if !item.body.trim().is_empty() {
-                        user_prompt.push_str("  Body:\n");
-                        user_prompt.push_str("  ");
-                        user_prompt.push_str(&item.body.replace('\n', "\n  "));
-                        user_prompt.push('\n');
-                    }
-                }
-            }
-            PrSummaryMode::ByPrs => {
-                let mut grouped: BTreeMap<u32, Vec<&PrItem>> = BTreeMap::new();
-                let mut no_pr: Vec<&PrItem> = Vec::new();
-
-                for item in items {
-                    if let Some(num) = item.pr_number {
-                        grouped.entry(num).or_default().push(item);
-                    } else {
-                        no_pr.push(item);
-                    }
-                }
-
-                user_prompt.push_str(
-                    "Pull requests contributing to this branch (oldest commits first):\n",
-                );
-
-                for (num, group) in grouped {
-                    let short = group[0]
-                        .commit_hash
-                        .chars()
-                        .take(7)
-                        .collect::<String>();
-                    let title = group[0].title.trim();
-                    user_prompt.push_str(&format!("\nPR #{num}: {title} [{short}]\n"));
-
-                    if group.len() > 1 {
-                        user_prompt.push_str("Additional commits in this PR:\n");
-                        for item in group.iter().skip(1) {
-                            let sh = item
-                                .commit_hash
-                                .chars()
-                                .take(7)
-                                .collect::<String>();
-                            user_prompt.push_str(&format!(
-                                "- {sh}: {title}\n",
-                                title = item.title.trim()
-                            ));
-                        }
-                    }
-                }
-
-                if !no_pr.is_empty() {
-                    user_prompt.push_str(
-                        "\nCommits without associated PR numbers (may be small fixes or direct pushes):\n",
-                    );
-                    for item in no_pr {
-                        let short = item
-                            .commit_hash
-                            .chars()
-                            .take(7)
-                            .collect::<String>();
-                        user_prompt.push_str(&format!(
-                            "- {short}: {title}\n",
-                            title = item.title.trim()
-                        ));
-                    }
-                }
-            }
-        }
-
-        if debug {
-            eprintln!(
-                "[DEBUG] PR description prompt:\n{}",
-                truncate(&user_prompt, 3500)
-            );
-        }
-
-        let req = ChatRequest {
-            model: self.model.clone(),
-            messages: vec![
-                ChatMessage {
-                    role: "system".into(),
-                    content: system_instructions,
-                },
-                ChatMessage {
-                    role: "user".into(),
-                    content: user_prompt,
-                },
-            ],
-        };
-
-        let (content, _usage) = self.call_chat(&req, debug)?;
-        Ok(content)
-    }
-}
-
-/// Truncate long strings for debug logging.
-fn truncate(s: &str, max_len: usize) -> String {
-    if s.len() <= max_len {
-        s.to_string()
-    } else {
-        format!("{}...\n[truncated {} chars]", &s[..max_len], s.len() - max_len)
-    }
 }
 
 /// Run a git command and capture stdout as String.
@@ -1012,15 +442,10 @@ fn run_pr(
 fn main() -> Result<()> {
     let cli = Cli::parse();
 
-    // Decide model + client
-    let model_from_cli = cli.model.clone();
-    let model_env = env::var("COMMITBOT_MODEL").ok();
-    let model = model_from_cli
-        .or(model_env)
-        .unwrap_or_else(|| "gpt-4o-mini".to_string());
+    // Resolve config (model from cli/env/toml)
+    let cfg = Config::from_sources(&cli);
 
-    let use_no_model = cli.no_model || model.to_lowercase() == "none";
-
+    let use_no_model = cli.no_model || cfg.model.to_lowercase() == "none";
     let api_key = cli.api_key.clone();
 
     let boxed_client: Box<dyn LlmClient> = if use_no_model {
@@ -1036,10 +461,10 @@ fn main() -> Result<()> {
         })?;
 
         if cli.debug {
-            eprintln!("[DEBUG] Using OpenAiClient with model: {model}");
+            eprintln!("[DEBUG] Using OpenAiClient with model: {}", cfg.model);
         }
 
-        Box::new(OpenAiClient::new(key, model))
+        Box::new(OpenAiClient::new(key, cfg.model.clone()))
     };
 
     match &cli.command {
